part 3
just finished the first part of my 2 touch throwing system. Now the
first touch defines the x position and the 2nd touch defines the 2nd
position.

This was tricky at first I need to make 3 bools. 2 to keep track of
which touch was just carried out and a 3rd to get the game to wait for
the next touch to carry out the both touches (that probably didn’t make
much sense to anyone but me, just like my variables)